Using float for Speed modifier

+++ TheCodeKitchen/TheCodeKitchen.Application.Contracts/Requests/CreateGameRequest.cs
@@ -1,4 +1,4 @@
 namespace TheCodeKitchen.Application.Contracts.Requests;
 
 [GenerateSerializer]
-public record CreateGameRequest(string? Name, double SpeedModifier);
+public record CreateGameRequest(string? Name, float SpeedModifier);
