+++ TheCodeKitchen/TheCodeKitchen.Application.Contracts/Requests/UpdateGameRequest.cs
@@ -1,4 +1,4 @@
 namespace TheCodeKitchen.Application.Contracts.Requests;
 
 [GenerateSerializer]
-public record UpdateGameRequest(double SpeedModifier);
+public record UpdateGameRequest(float SpeedModifier);
